Ajout des formules de calcul. Prise en compte effective de la classe State dans l'exécution du programme. Ajout du levier dotation récurrente à la recherche qui avait été oublié

See merge request t3-2019-2020-derousseaux-van-der-spurt-treyer/t432_haj19_t3_a!29

+++ New Presidency/src/model/LeverList.java
@@ -26,7 +26,7 @@ public class LeverList {
         _levers=new ArrayList<>();
     }
 
-    public Lever createLever(String name, Integer initBudget, ArrayList<String> infos){
+    public Lever createLever(String name, Double initBudget, ArrayList<String> infos){
         Lever l = new Lever(name,initBudget,infos);
         _levers.add(l);
         return l;
